+++ CHANGELOG.md
@@ -86,6 +86,8 @@
 * Sorting on the Instance's Holdings item table is not working, part 2. Fixes UIIN-2565.
 * Promote a local instance to become a shared instance. Refs UIIN-2460.
 * Consortial Central Tenant: Suppress Holdings detail view Action menu. Refs UIIN-2524.
+* Consortial Central Tenant: Suppress Item detail view Action menu. Refs UIIN-2525.
+* Consortial Central Tenant: Hide Instance Action Menu New Request option. Refs UIIN-2572.
 * Add facet for members with holdings on Instances in Inventory Instances/Holdings/Items search. Refs UIIN-2395.
 
 ## [9.4.11](https://github.com/folio-org/ui-inventory/tree/v9.4.11) (2023-08-02)